[refactor]:Main #35 Team package 생성 및 파일 이동
