+++ MainService/src/main/java/com/kusitms/mainservice/domain/team/repository/TeamRepository.java
@@ -1,6 +1,6 @@
-package com.kusitms.mainservice.domain.user.repository;
+package com.kusitms.mainservice.domain.team.repository;
 
-import com.kusitms.mainservice.domain.user.domain.Team;
+import com.kusitms.mainservice.domain.team.domain.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamRepository extends JpaRepository<Team, Long> {
